+++ benchmarks/benchmarks/compute.py
@@ -2,7 +2,7 @@
 import timeit
 import copy
 
-from pyphi import compute, config, constants, examples, Subsystem
+from pyphi import cache as _cache, compute, config, constants, examples, Subsystem
 from .subsystem import clear_subsystem_caches
 
 
@@ -12,26 +12,36 @@ def _clear_joblib_cache():
 
 class BenchmarkConstellation:
 
-    params = ['basic', 'rule154', 'fig16']
+    params = [
+        ['parallel', 'sequential'],
+        ['basic', 'rule154', 'fig16']
+    ]
+    param_names = ['mode', 'network']
+    number = 1
+    repeat = 1
 
-    def setup(self, network):
+    def setup(self, mode, network):
         if network == 'basic':
             self.subsys = examples.basic_subsystem()
-
         elif network == 'rule154':
             network = examples.rule154_network()
             state = (1,) * 5
             self.subsys = Subsystem(network, state, network.node_indices)
-
         elif network == 'fig16':
             network = examples.fig16()
             state = (0,) * 7
             self.subsys = Subsystem(network, state, network.node_indices)
-
         else:
-            raise
+            raise ValueError(network)
 
-    def time_constellation(self, network):
+        if mode == 'parallel':
+            config.PARALLEL_CONCEPT_EVALUATION = True
+        elif mode == 'sequential':
+            config.PARALLEL_CONCEPT_EVALUATION = False
+        else:
+            raise ValueError(mode)
+
+    def time_constellation(self, mode, network):
         clear_subsystem_caches(self.subsys)
         # network purview caches are left intact
         compute.constellation(self.subsys)
@@ -41,9 +51,10 @@ class BenchmarkMainComplex():
 
     params = [
         ['parallel', 'sequential'],
-        ['basic', 'rule154', 'fig16']
+        ['basic', 'rule154', 'fig16'],
+        ['local', 'redis']
     ]
-    param_names = ['mode', 'network']
+    param_names = ['mode', 'network', 'cache']
 
     # Use `default_timer` (clock time) instead of process time because
     # parallel execution spawns separate processes which are not counted
@@ -56,7 +67,7 @@ class BenchmarkMainComplex():
     repeat = 1
     timeout = 10000
 
-    def setup(self, mode, network):
+    def setup(self, mode, network, cache):
 
         if network == 'basic':
             self.network = examples.basic_network()
@@ -68,7 +79,7 @@ class BenchmarkMainComplex():
             self.network = examples.fig16()
             self.state = (1, 0, 0, 1, 1, 1, 0)
         else:
-            raise
+            raise ValueError(network)
 
         # Save config
         self.default_config = copy.copy(config.__dict__)
@@ -79,14 +90,25 @@ class BenchmarkMainComplex():
         elif mode == 'sequential':
             config.PARALLEL_CUT_EVALUATION = False
         else:
-            raise
+            raise ValueError(mode)
+
+        # Cache mode
+        if cache == 'local':
+            config.REDIS_CACHE = False
+        elif cache == 'redis':
+            config.REDIS_CACHE = True
+            if _cache.RedisConn().ping() is False:
+                # No server running
+                raise NotImplementedError
+        else:
+            raise ValueError(cache)
 
         config.CACHE_BIGMIPS = False
 
-    def teardown(self, mode, network):
+    def teardown(self, mode, network, cache):
         # Revert config
         config.__dict__.update(self.default_config)
 
-    def time_main_complex(self, mode, network):
+    def time_main_complex(self, mode, network, cache):
         # Do it!
         compute.main_complex(self.network, self.state)