+++ test/test_models.py
@@ -5,6 +5,7 @@
 from unittest import mock
 from collections import namedtuple
 import numpy as np
+import pytest
 
 from pyphi import models, constants, config, Subsystem
 
@@ -21,93 +22,117 @@ a = nt(this=('consciousness', 'is phi'), that=np.arange(3), phi=0.5,
 # Test equality helpers {{{
 # =========================
 
-def test_phi_comparisons():
+def test_phi_mechanism_ordering():
 
-    class PhiThing:
-        def __init__(self, phi):
+    class PhiThing(models.cmp._Orderable):
+        def __init__(self, phi, mechanism):
             self.phi = phi
-    small = PhiThing(0.0)
-    large = PhiThing(2.0)
+            self.mechanism = mechanism
 
-    assert models._phi_lt(small, large)
-    assert not models._phi_lt(large, small)
-    assert not models._phi_lt(small, small)
-    assert not models._phi_lt(small, 'attr_error')
+        def _order_by(self):
+            return [self.phi, self.mechanism]
 
-    assert models._phi_le(small, large)
-    assert not models._phi_le(large, small)
-    assert models._phi_le(small, small)
-    assert not models._phi_le(small, 'attr_error')
+        def __eq__(self, other):
+            return self.phi == other.phi and self.mechanism == other.mechanism
 
-    assert models._phi_gt(large, small)
-    assert not models._phi_gt(small, large)
-    assert not models._phi_gt(large, large)
-    assert not models._phi_gt(small, 'attr_error')
+    # assert PhiThing(1.0, (1,)) == PhiThing(1.0, (1,))
+    # assert PhiThing(1.0, (1,)) == PhiThing(1.0, (1, 2))
+    # assert PhiThing(1.0, (1,)) != PhiThing(2.0, (1, 2))
+    assert PhiThing(1.0, (1,)) < PhiThing(2.0, (1,))
+    assert PhiThing(1.0, (1,)) <= PhiThing(1.0, (1, 2))  # Smaller mechanism
+    assert PhiThing(1.0, (1,)) <= PhiThing(2.0, (1,))
+    assert PhiThing(2.0, (1,)) > PhiThing(1.0, (1,))
+    assert PhiThing(2.0, (1,)) > PhiThing(1.0, (1, 2))  # Larger phi
+    assert PhiThing(1.0, (1,)) >= PhiThing(1.0, (1,))
+    assert PhiThing(1.0, (1, 2)) >= PhiThing(1.0, (1,))
 
-    assert models._phi_ge(large, small)
-    assert not models._phi_ge(small, large)
-    assert models._phi_ge(large, large)
-    assert not models._phi_ge(small, 'attr_error')
+    class PhiLikeThing(PhiThing):
+        pass
+
+    # Compared objects must be of the same class
+    with pytest.raises(TypeError):  # TypeError: unorderable types
+        PhiThing(1.0, (1, 2)) <= PhiLikeThing(1.0, (1, 2))
+
+    class PhiThang(PhiThing):
+        def __init__(self, phi, mechanism, purview):
+            super().__init__(phi, mechanism)
+            self.purview = purview
+
+        def __eq__(self, other):
+            return self.purview == other.purview
+
+    assert PhiThang(1.0, (1,), (1,)) == PhiThang(2.0, (3,), (1,))
+    assert PhiThang(1.0, (1,), (1,)) < PhiThang(2.0, (1,), (2,))
+
+
+def test_sametype_decorator():
+
+    class Thing:
+        @models.cmp.sametype
+        def do_it(self, other):
+            return True
+
+    assert Thing().do_it(object()) == NotImplemented
 
 
 def test_numpy_aware_eq_noniterable():
     b = 1
-    assert not models._numpy_aware_eq(a, b)
+    assert not models.cmp._numpy_aware_eq(a, b)
 
 
 def test_numpy_aware_eq_nparray():
     b = np.arange(3)
-    assert not models._numpy_aware_eq(a, b)
+    assert not models.cmp._numpy_aware_eq(a, b)
 
 
 def test_numpy_aware_eq_tuple_nparrays():
     b = (np.arange(3), np.arange(3))
-    assert not models._numpy_aware_eq(a, b)
+    assert not models.cmp._numpy_aware_eq(a, b)
 
 
 def test_numpy_aware_eq_identical():
     b = a
-    assert models._numpy_aware_eq(a, b)
+    assert models.cmp._numpy_aware_eq(a, b)
 
 
 def test_general_eq_different_attributes():
     similar_nt = namedtuple('nt', nt_attributes + ['supbro'])
     b = similar_nt(a.this, a.that, a.phi, a.mechanism, a.purview,
                    supbro="nothin' much")
-    assert models._general_eq(a, b, nt_attributes)
+    assert models.cmp._general_eq(a, b, nt_attributes)
 
 
 def test_general_eq_phi_precision_comparison_true():
     b = nt(a.this, a.that, (a.phi - constants.EPSILON/2), a.mechanism,
            a.purview)
-    assert models._general_eq(a, b, nt_attributes)
+    assert models.cmp._general_eq(a, b, nt_attributes)
 
 
 def test_general_eq_phi_precision_comparison_false():
     b = nt(a.this, a.that, (a.phi - constants.EPSILON*2), a.mechanism,
            a.purview)
-    assert not models._general_eq(a, b, nt_attributes)
+    assert not models.cmp._general_eq(a, b, nt_attributes)
 
 
 def test_general_eq_different_mechanism_order():
     b = nt(a.this, a.that, a.phi, a.mechanism[::-1], a.purview)
-    assert models._general_eq(a, b, nt_attributes)
+    assert models.cmp._general_eq(a, b, nt_attributes)
 
 
 def test_general_eq_different_purview_order():
     b = nt(a.this, a.that, a.phi, a.mechanism, a.purview[::-1])
-    assert models._general_eq(a, b, nt_attributes)
+    assert models.cmp._general_eq(a, b, nt_attributes)
 
 
 def test_general_eq_different_mechanism_and_purview_order():
     b = nt(a.this, a.that, a.phi, a.mechanism[::-1], a.purview[::-1])
-    assert models._general_eq(a, b, nt_attributes)
+    assert models.cmp._general_eq(a, b, nt_attributes)
 
 
 # }}}
 
-# Test Cut
-# ========
+# Test Cut {{{
+# ============
 
 def test_cut_splits_mechanism():
     cut = models.Cut((0,), (1, 2))
@@ -147,53 +172,45 @@ def test_cut_matrix():
     assert np.array_equal(cut.cut_matrix(), np.array([]))
 
 
+# }}}
+
+def mip(phi=1.0, dir=None, mech=(), purv=(), partition=None,
+        unpartitioned_repertoire=None, partitioned_repertoire=None):
+    return models.Mip(phi=phi, direction=dir, mechanism=mech,
+                      purview=purv, partition=partition,
+                      unpartitioned_repertoire=unpartitioned_repertoire,
+                      partitioned_repertoire=partitioned_repertoire)
+
+
 # Test MIP {{{
 # ============
 
-def test_mip_ordering():
-    phi1 = models.Mip(
-        direction=None, mechanism=(), purview=(), partition=None,
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=1.0)
-    different_phi1 = models.Mip(
-        direction='different', mechanism=(), purview=(), partition=0,
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=1.0)
-    phi2 = models.Mip(
-        direction=0, mechanism=(), purview=(), partition='stilldifferent',
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=1.0 + constants.EPSILON*2)
-    assert phi1 < phi2
-    assert phi2 > phi1
-    assert phi1 <= phi2
-    assert phi2 >= phi1
-    assert phi1 <= different_phi1
-    assert phi1 >= different_phi1
+def test_mip_ordering_and_equality():
+    assert mip(phi=1.0) < mip(phi=2.0)
+    assert mip(phi=2.0) > mip(phi=1.0)
+    assert mip(phi=1.0, mech=(1,)) < mip(phi=1.0, mech=(1, 2))
+    assert mip(phi=1.0, mech=(1, 2)) >= mip(phi=1.0, mech=(1,))
+    assert mip(phi=1.0, mech=(1,), purv=(1,)) < mip(phi=1.0, mech=(1,), purv=(1, 2))
+    assert mip(phi=1.0, mech=(1,), purv=(1, 2)) >= mip(phi=1.0, mech=(1,), purv=(1,))
 
+    assert mip(phi=1.0) == mip(phi=1.0)
+    assert mip(phi=1.0) == mip(phi=(1.0 - constants.EPSILON/2))
+    assert mip(phi=1.0) != mip(phi=(1.0 - constants.EPSILON * 2))
+    assert mip(dir='past') != mip(dir='future')
+    assert mip(mech=(1,)) != mip(mech=(1, 2))
 
-def test_mip_equality():
-    phi = 1.0
-    mip = models.Mip(
-        direction=None, mechanism=(), purview=(), partition=None,
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=phi)
-    close_enough = models.Mip(
-        direction=None, mechanism=(), purview=(), partition=None,
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=(phi - constants.EPSILON/2))
-    not_quite = models.Mip(
-        direction=None, mechanism=(), purview=(), partition=None,
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=(phi - constants.EPSILON*2))
-    assert mip == close_enough
-    assert mip != not_quite
+    with pytest.raises(TypeError):
+        mip(dir='past') < mip(dir='future')
+
+    with pytest.raises(TypeError):
+        mip(dir='past') >= mip(dir='future')
 
 
 def test_null_mip():
     direction = 'past'
     mechanism = (0,)
     purview = (1,)
-    null_mip = models.Mip._null_mip(direction, mechanism, purview)
+    null_mip = models._null_mip(direction, mechanism, purview)
     assert null_mip.direction == direction
     assert null_mip.mechanism == mechanism
     assert null_mip.purview == purview
@@ -218,35 +235,24 @@ def test_mip_repr_str():
 # =============
 
 def test_mice_ordering_by_phi():
-    phi1 = models.Mice(models.Mip(
-        direction=None, mechanism=(), purview=(),
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=1.0, partition=()))
-    different_phi1 = models.Mice(models.Mip(
-        direction='different', mechanism=(), purview=(),
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=1.0, partition=()))
-    phi2 = models.Mice(models.Mip(
-        direction=None, mechanism=(), purview=(),
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=(1.0 + constants.EPSILON*2), partition=()))
+    phi1 = models.Mice(mip())
+    different_phi1 = models.Mice(mip(dir='different'))
+    phi2 = models.Mice(mip(phi=(1.0 + constants.EPSILON * 2), partition=()))
     assert phi1 < phi2
     assert phi2 > phi1
     assert phi1 <= phi2
     assert phi2 >= phi1
-    assert phi1 <= different_phi1
-    assert phi1 >= different_phi1
+
+    with pytest.raises(TypeError):
+        phi1 <= different_phi1
+
+    with pytest.raises(TypeError):
+        phi1 >= different_phi1
 
 
 def test_mice_odering_by_mechanism():
-    small = models.Mice(models.Mip(
-        direction=None, mechanism=(1, 2), purview=(),
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=1.0, partition=()))
-    big = models.Mice(models.Mip(
-        direction=None, mechanism=(1, 2, 3), purview=(),
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=1.0, partition=()))
+    small = models.Mice(mip(mech=(1,)))
+    big = models.Mice(mip(mech=(1, 2, 3)))
     assert small < big
     assert small <= big
     assert big > small
@@ -254,20 +260,19 @@ def test_mice_odering_by_mechanism():
     assert big != small
 
 
+def test_mice_ordering_by_purview():
+    small = models.Mice(mip(purv=(1, 2)))
+    big = models.Mice(mip(purv=(1, 2, 3)))
+    assert small < big
+    assert small <= big
+    assert big > small
+    assert big >= small
+
+
 def test_mice_equality():
-    phi = 1.0
-    mice = models.Mice(models.Mip(
-        direction=None, mechanism=(), purview=(),
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=phi, partition=()))
-    close_enough = models.Mice(models.Mip(
-        direction=None, mechanism=(), purview=(),
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=(phi - constants.EPSILON/2), partition=()))
-    not_quite = models.Mice(models.Mip(
-        direction=None, mechanism=(), purview=(),
-        unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=(phi - constants.EPSILON*2), partition=()))
+    mice = models.Mice(mip(phi=1.0))
+    close_enough = models.Mice(mip(phi=(1.0 - constants.EPSILON / 2)))
+    not_quite = models.Mice(mip(phi=(1.0 - constants.EPSILON * 2)))
     assert mice == close_enough
     assert mice != not_quite
 
@@ -317,35 +322,39 @@ def test_damaged(s):
     assert mice.damaged_by_cut(subsys)
     assert not mice.damaged_by_cut(s)
 
+
 # }}}
 
 # Test Concept {{{
 # ================
 
-def test_concept_ordering():
+def test_concept_ordering(s, micro_s):
     phi1 = models.Concept(
-        mechanism=(0, 1), cause=1, effect=None, subsystem=None,
+        mechanism=(0, 1), cause=1, effect=None, subsystem=s,
         phi=1.0)
     different_phi1 = models.Concept(
-        mechanism=(), cause='different', effect=None, subsystem=None,
+        mechanism=(0, 1), cause='different', effect=None, subsystem=micro_s,
         phi=1.0)
     phi2 = models.Concept(
-        mechanism=0, cause='stilldifferent', effect=None, subsystem=None,
+        mechanism=(0,), cause='stilldifferent', effect=None, subsystem=s,
         phi=1.0 + constants.EPSILON*2)
     assert phi1 < phi2
     assert phi2 > phi1
     assert phi1 <= phi2
     assert phi2 >= phi1
-    assert phi1 <= different_phi1
-    assert phi1 >= different_phi1
+
+    with pytest.raises(TypeError):
+        phi1 <= different_phi1
+    with pytest.raises(TypeError):
+        phi1 > different_phi1
 
 
-def test_concept_odering_by_mechanism():
+def test_concept_odering_by_mechanism(s):
     small = models.Concept(
-        mechanism=(0, 1), cause=None, effect=None, subsystem=None,
+        mechanism=(0, 1), cause=None, effect=None, subsystem=s,
         phi=1.0)
     big = models.Concept(
-        mechanism=(0, 1, 3), cause=None, effect=None, subsystem=None,
+        mechanism=(0, 1, 3), cause=None, effect=None, subsystem=s,
         phi=1.0)
     assert small < big
     assert small <= big
@@ -354,12 +363,102 @@ def test_concept_odering_by_mechanism():
     assert big != small
 
 
-def test_concept_equality():
+def test_concept_equality(s):
     phi = 1.0
-    concept = models.Concept(
-        mechanism=(), cause=None, effect=None, subsystem=None, phi=phi)
-    another = models.Concept(
-        mechanism=(), cause=None, effect=None, subsystem=None, phi=phi)
+    concept = models.Concept(mechanism=(), cause=None, effect=None,
+                             subsystem=s, phi=phi)
+    another = models.Concept(mechanism=(), cause=None, effect=None,
+                             subsystem=s, phi=phi)
+    assert concept == another
+
+
+def test_concept_equality_phi(s):
+    concept = models.Concept(mechanism=(), cause=None, effect=None,
+                             subsystem=s, phi=1.0)
+    another = models.Concept(mechanism=(), cause=None, effect=None,
+                             subsystem=s, phi=0.0)
+    assert concept != another
+
+
+def test_concept_equality_mechanism(s):
+    phi = 1.0
+    concept = models.Concept(mechanism=(1,), cause=None, effect=None,
+                             subsystem=s, phi=phi)
+    another = models.Concept(mechanism=(), cause=None, effect=None,
+                             subsystem=s, phi=phi)
+    assert concept != another
+
+
+def test_concept_equality_cause_purview_nodes(s):
+    phi = 1.0
+    mice1 = models.Mice(models.Mip(
+        direction=None, mechanism=(), purview=(1, 2),
+        unpartitioned_repertoire=None, partitioned_repertoire=None,
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
+    mice2 = models.Mice(models.Mip(
+        direction=None, mechanism=(), purview=(1,),
+        unpartitioned_repertoire=None, partitioned_repertoire=None,
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
+    concept = models.Concept(mechanism=(), cause=mice1, effect=None,
+                             subsystem=s, phi=phi)
+    another = models.Concept(mechanism=(), cause=mice2, effect=None,
+                             subsystem=s, phi=phi)
+    assert concept != another
+
+
+def test_concept_equality_effect_purview_nodes(s):
+    phi = 1.0
+    mice1 = models.Mice(models.Mip(
+        direction=None, mechanism=(), purview=(1, 2),
+        unpartitioned_repertoire=None, partitioned_repertoire=None,
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
+    mice2 = models.Mice(models.Mip(
+        direction=None, mechanism=(), purview=(1,),
+        unpartitioned_repertoire=None, partitioned_repertoire=None,
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
+    concept = models.Concept(mechanism=(), cause=None, effect=mice1,
+                             subsystem=s, phi=phi)
+    another = models.Concept(mechanism=(), cause=None, effect=mice2,
+                             subsystem=s, phi=phi)
+    assert concept != another
+
+
+def test_concept_equality_repertoires(s):
+    phi = 1.0
+    mice1 = models.Mice(models.Mip(
+        direction=None, mechanism=(), purview=(),
+        unpartitioned_repertoire=np.array([1, 2]), partitioned_repertoire=(),
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
+    mice2 = models.Mice(models.Mip(
+        direction=None, mechanism=(), purview=(),
+        unpartitioned_repertoire=np.array([0, 0]), partitioned_repertoire=None,
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
+    concept = models.Concept(mechanism=(), cause=mice1, effect=mice2,
+                             subsystem=s, phi=phi)
+    another = models.Concept(mechanism=(), cause=mice2, effect=mice1,
+                             subsystem=s, phi=phi)
+    assert concept != another
+
+
+def test_concept_equality_network(s, simple_subsys_all_off):
+    phi = 1.0
+    concept = models.Concept(mechanism=(), cause=None, effect=None,
+                             subsystem=simple_subsys_all_off, phi=phi)
+    another = models.Concept(mechanism=(), cause=None, effect=None,
+                             subsystem=s, phi=phi)
+    assert concept != another
+
+
+def test_concept_equality_one_subsystem_is_subset_of_another(s, subsys_n1n2):
+    phi = 1.0
+    mice = models.Mice(models.Mip(
+        direction=None, mechanism=(), purview=(1, 2),
+        unpartitioned_repertoire=(), partitioned_repertoire=(),
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
+    concept = models.Concept(mechanism=(2,), cause=mice, effect=mice,
+                             subsystem=s, phi=phi)
+    another = models.Concept(mechanism=(2,), cause=mice, effect=mice,
+                             subsystem=subsys_n1n2, phi=phi)
     assert concept == another
 
 
@@ -375,20 +474,34 @@ def test_concept_repr_str():
     print(str(concept))
 
 
-def test_concept_hashing():
+def test_concept_hashing(s):
     mice = models.Mice(models.Mip(
         direction=None, mechanism=(0, 1, 2), purview=(0, 1, 2),
         unpartitioned_repertoire=None, partitioned_repertoire=None,
         phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
     concept = models.Concept(
-        mechanism=(0, 1, 2), cause=mice, effect=mice, subsystem=None, phi=0.0)
+        mechanism=(0, 1, 2), cause=mice, effect=mice, subsystem=s, phi=0.0)
     hash(concept)
+
+
+def test_concept_hashing_one_subsystem_is_subset_of_another(s, subsys_n1n2):
+    phi = 1.0
+    mice = models.Mice(models.Mip(
+        direction=None, mechanism=(), purview=(1, 2),
+        unpartitioned_repertoire=(), partitioned_repertoire=(),
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
+    concept = models.Concept(mechanism=(2,), cause=mice, effect=mice,
+                             subsystem=s, phi=phi)
+    another = models.Concept(mechanism=(2,), cause=mice, effect=mice,
+                             subsystem=subsys_n1n2, phi=phi)
+    assert hash(concept) == hash(another)
+    assert(len(set([concept, another])) == 1)
+
 
 # }}}
 
-
-# Test Constellation
-# ==================
+# Test Constellation {{{
+# ======================
 
 def test_constellation_is_still_a_tuple():
     c = models.Constellation([models.Concept()])
@@ -401,38 +514,42 @@ def test_constellation_repr():
     assert repr(c) == "Constellation(())"
 
 
+# }}}
+
 # Test BigMip {{{
 # ===============
 
-def test_bigmip_ordering():
+def test_bigmip_ordering(s, s_noised):
     phi1 = models.BigMip(
         unpartitioned_constellation=None, partitioned_constellation=None,
-        subsystem=(), cut_subsystem=(),
+        subsystem=s, cut_subsystem=(),
         phi=1.0)
     different_phi1 = models.BigMip(
         unpartitioned_constellation=0, partitioned_constellation=None,
-        subsystem=(), cut_subsystem=(),
+        subsystem=s_noised, cut_subsystem=(),
         phi=1.0)
     phi2 = models.BigMip(
         unpartitioned_constellation='stilldifferent',
-        partitioned_constellation=None, subsystem=(), cut_subsystem=(),
+        partitioned_constellation=None, subsystem=s, cut_subsystem=(),
         phi=1.0 + constants.EPSILON*2)
     assert phi1 < phi2
     assert phi2 > phi1
     assert phi1 <= phi2
     assert phi2 >= phi1
-    assert phi1 <= different_phi1
-    assert phi1 >= different_phi1
+    with pytest.raises(TypeError):
+        phi1 <= different_phi1
+    with pytest.raises(TypeError):
+        phi1 >= different_phi1
 
 
-def test_bigmip_odering_by_mechanism():
+def test_bigmip_ordering_by_subsystem_size(s, s_single):
     small = models.BigMip(
         unpartitioned_constellation=None,
-        partitioned_constellation=None, subsystem=[1], cut_subsystem=(),
+        partitioned_constellation=None, subsystem=s_single, cut_subsystem=(),
         phi=1.0)
     big = models.BigMip(
         unpartitioned_constellation=None,
-        partitioned_constellation=None, subsystem=[1, 2], cut_subsystem=(),
+        partitioned_constellation=None, subsystem=s, cut_subsystem=(),
         phi=1.0)
     assert small < big
     assert small <= big
@@ -469,15 +586,15 @@ def test_bigmip_repr_str(s):
 
 # }}}
 
-
-# Test model __str__ and __reprs__
+# Test model __str__ and __reprs__ {{{
+# ====================================
 
 def test_indent():
     s = ("line1\n"
          "line2")
     answer = ("  line1\n"
               "  line2")
-    assert models.indent(s) == answer
+    assert models.fmt.indent(s) == answer
 
 
 class ReadableReprClass:
@@ -485,7 +602,7 @@ class ReadableReprClass:
     some_attr = 3.14
 
     def __repr__(self):
-        return models.make_repr(self, ['some_attr'])
+        return models.fmt.make_repr(self, ['some_attr'])
 
     def __str__(self):
         return "A nice fat explicit string"
@@ -500,5 +617,6 @@ def test_make_reprs_uses___repr__():
 def test_make_reprs_calls_out_to_string():
     assert repr(ReadableReprClass()) == "A nice fat explicit string"
 
+# }}}
 
 # vim: set foldmarker={{{,}}} foldlevel=0  foldmethod=marker :